Replacing CLS with local-storage, simplified usage which should remove the memory leak permenantly.
