+++ packages/backend-core/db.js
@@ -1,6 +1,1 @@
-module.exports = {
-  ...require("./src/db/utils"),
-  ...require("./src/db/constants"),
-  ...require("./src/db"),
-  ...require("./src/db/views"),
-}
+module.exports = require("./src/db")
