+++ packages/backend-core/src/context/constants.ts
@@ -1,9 +1,5 @@
 import { IdentityContext } from "@budibase/types"
 
-export enum ContextKey {
-  MAIN = "main",
-}
-
 export type ContextMap = {
   tenantId?: string
   appId?: string
