+++ packages/backend-core/src/index.ts
@@ -24,7 +24,7 @@ import * as queue from "./queue"
 import * as types from "./types"
 
 // mimic the outer package exports
-import * as db from "./pkg/db"
+import * as db from "./db"
 import * as objectStore from "./pkg/objectStore"
 import * as utils from "./pkg/utils"
 import redis from "./pkg/redis"
