--- packages/backend-core/src/pkg/db.ts
@@ -1,6 +1,0 @@
-// Mimic the outer package export for usage in index.ts
-// The outer exports can't be used as they now reference dist directly
-export * from "../db"
-export * from "../db/utils"
-export * from "../db/views"
-export * from "../db/constants"
